Validate chunk headers on SerializationQueue thread
Validating the chunk headers on the SerializationQueue thread
alleviates the race condition where a new channel security is being
installed but the header validation uses the old channel security.
Installation happens on the SerializationQueue, so moving validation to
the queue eliminates the possibility that installation and validation
happen concurrently.

Fixes issues #8