Copy chunkBuffers in a thread-safe manner before decoding
It’s difficult to reason about whether the thread that ends up running
decode logic on the serialization queue is guaranteed to see all
updates to the buffer list or if it’s only guaranteed to see the
reference to it. Making a copy using a thread-safe immutable collection
eliminates any uncertainty.